pyproject.toml
[project]
name = "knowledge-base"
version = "0.1.0"
description = "Agentic Knowledge Ingestion & Management System"
readme = "README.md"
authors = [
    { name = "monjurkuet", email = "muhammad.manjur@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.23",
    "pgvector>=0.2.4",
    "google-generativeai>=0.3.0",
    "openai>=1.3.0",
    "logfire>=0.28.0",
    "unstructured[md]>=0.11.0",
    "igraph>=0.11.0",
    "leidenalg>=0.10.0",
    "python-dateutil>=2.8.2",
    "dateparser>=1.2.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "fastapi>=0.128.0",
    "pytest-asyncio>=1.3.0",
    "psycopg2-binary>=2.9.11",
    "sqlparse>=0.5.5",
    "uvicorn[standard]>=0.40.0",
    "rich>=13.0",
    "nltk>=3.8.0",
    "websockets>=16.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2.2",
    "tenacity>=8.0.0",
]

[project.scripts]
knowledge-base = "knowledge_base:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
